Added files to test the Nqubit controlled NOT and UNITARY member functions of the Qubit_Extended class.
Tests work as expected.

Note there exists a CHEAT version which needs to be updated to work which uses only
two qubits, but will peak at the states' amplitudes. This would not be possible for a real quntum systyem and so should
not be assumed. However, if scaling becomes an issue, it might be useful to complete this member function if the results are
shown to be the same, so that larger scale prblems can be explored.