Modified to correctly measure multiple qubit in a state simultaneously. This is done from sampling according to the distribution of the probability of each state being observed.
Further work needs to be done to validate. There exists an error which causes
the run of each experiment being not quite as expected.

Another issue which needs to be resolved is that there is a strong bias towards
assigning a class which has multiple element states. Further investigation required.