Updated the IntelSimulator.cpp class to incorporate measurement methods. These methods still have to be implemented in the SimulatorGeneral class and file.
Also note that there is an issue with a shared object not being found at
runtime of the exe_NCU executabgle built in demos using the Simulator
class. Using otool -L to see the linked dependencies it is visible as
beig linked. This is strange behaviour and required further
investigation.